Fixed a bug with the translate command
Unlike, the translate command in Processing, javascript command makes
a permanent change to the context of the canvas that persists between
draw cycles. To correct this behavior, I had to save the current
context at the start of the draw cycle and restore at the end to make
sure that each new draw gets a fresh context.